requirements.txt
face_recognition==1.3.0
matplotlib==3.8.4
numpy==2.0.1
Pillow==10.4.0
pyserial==3.5
torch==2.2.0
torchvision==0.15.2a0
transformers==4.41.2
opencv==4.7.0
